refactor: update model imports to lowercase and create new post and user model files
